refactors tests for easier translation groking
