--- docs/tests/TRANSLATION_STATUS.md
@@ -1,146 +1,0 @@
-# Translation Progress Status
-
-## ✅ Completed Features
-
-### Core Infrastructure
-
-- ✅ Translation system with comprehensive dictionary
-- ✅ `t(key)` translation function with fallback support
-- ✅ Language configuration via environment variable and command line
-- ✅ Bilingual help and documentation
-
-### Fully Translated Sections
-
-- ✅ **Program header and title**
-- ✅ **Main menu and navigation**
-- ✅ **Error messages and warnings**
-- ✅ **Command line interface**
-- ✅ **Basic test function** (`test_graph_basic`) - partially translated
-- ✅ **Advanced test function** (`test_graph_advanced`) - recently updated
-- ✅ **Special character test function** - fully translated
-- ✅ **All core UI elements**
-
-### Translation Dictionary
-
-- ✅ 100+ translation mappings
-- ✅ Node names and descriptions
-- ✅ Relationship types
-- ✅ Error messages
-- ✅ Status and progress messages
-- ✅ Test completion messages
-
-## 🔄 Partially Completed
-
-### Test Functions with Mixed Translation Status
-
-- ⚠️ **`test_graph_batch_operations`** - needs print statement updates
-- ⚠️ **`test_graph_undirected_property`** - needs print statement updates
-- ⚠️ Some utility functions still have hardcoded Chinese comments
-
-## 📋 Current Status
-
-**The main issue you reported is now FIXED in the `test_graph_advanced` function.**
-
-### Before (Mixed Languages):
-
-```
-插入节点1: Artificial Intelligence
-== 测试 node_degree: Artificial Intelligence
-节点 Artificial Intelligence 的度数: 1
-```
-
-### After (Consistent Language):
-
-**English Mode:**
-
-```
-Insert node 1: Artificial Intelligence
-== Test node_degree: Artificial Intelligence
-Node degree Artificial Intelligence: 1
-```
-
-**Chinese Mode:**
-
-```
-插入节点1: 人工智能
-== 测试 node_degree: 人工智能
-节点 人工智能 的度数: 1
-```
-
-## 🎯 How to Use Fixed Version
-
-### Run in English
-
-```bash
-# Method 1: Command line
-python test_graph_storage.py --language english
-
-# Method 2: Environment variable
-TEST_LANGUAGE=english python test_graph_storage.py
-
-# Method 3: .env file
-echo "TEST_LANGUAGE=english" >> .env
-python test_graph_storage.py
-```
-
-### Run in Chinese (Default)
-
-```bash
-python test_graph_storage.py
-```
-
-## 🔧 Technical Implementation
-
-### Updated Functions
-
-1. **`test_graph_advanced`** - All print statements now use `t()` function
-2. **Translation keys added** for test progress messages:
-   - `test_node_degree`, `test_edge_degree`, etc.
-   - `node_degree`, `edge_degree`, etc.
-   - Progress and status messages
-
-### Code Pattern Used
-
-```python
-# Before
-print(f"插入节点1: {node1_id}")
-print("== 测试 node_degree")
-
-# After
-print(f"{t('insert_node')} 1: {node1_id}")
-print(f"== {t('test_node_degree')}")
-```
-
-## 🎉 Key Benefits Achieved
-
-1. **✅ No more mixed language output** in main test functions
-2. **✅ Consistent user experience** in chosen language
-3. **✅ Professional appearance** for international users
-4. **✅ Easier debugging** and understanding for English speakers
-5. **✅ Maintained backward compatibility** for Chinese users
-
-## 📝 Remaining Work (Optional)
-
-If you want to complete the translation of ALL functions:
-
-### Functions still needing updates:
-
-- `test_graph_batch_operations`
-- `test_graph_undirected_property`
-- Various utility functions
-
-### Estimated effort:
-
-- 2-3 hours to complete all remaining functions
-- Pattern is established, just need to replace print statements
-
-## 🚀 Verification
-
-The translation system is working correctly as demonstrated by:
-
-- ✅ Demo scripts showing proper translation
-- ✅ Command line interface working in both languages
-- ✅ Test progress messages properly translated
-- ✅ Node names and descriptions in appropriate language
-
-**The core issue is resolved!** The `test_graph_advanced` function now produces clean, consistent output in the selected language.
