+++ tests/test_graph_storage/__init__.py
@@ -1,0 +1,7 @@
+"""
+Test Graph Storage Package
+Organized test suite for graph storage implementations
+"""
+
+__version__ = "1.0.0"
+__author__ = "LightRAG Team"
