+++ tests/test_graph_storage/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Core utilities for graph storage testing
+"""
