+++ tests/test_graph_storage/tests/__init__.py
@@ -1,0 +1,6 @@
+"""
+Test modules for graph storage implementations
+
+This module serves as the entry point for test modules.
+All utility functions are imported from core/storage_setup.py to avoid duplication.
+"""
